requirements.txt
Flask
flask_jwt_extended
flask-migrate
Flask-SQLAlchemy
flask-swagger-ui
gunicorn 
PyMySQL
pytest
python-dotenv